requirements.txt
Django
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0